+++ src/nodes/s2-resource/index.ts
@@ -374,6 +374,7 @@ export = function (RED: NodeRedApp): void {
           tcpAddress: dbusConfig.tcpAddress,
           deviceType: dbusConfig.deviceType,
           nodeId: node.id,
+          customName: config.rmName || config.name,
           measurementType: dbusConfig.measurementType,
           nrOfPhases: dbusConfig.nrOfPhases,
           position: dbusConfig.position,